refactor: PathFinder인터페이스를 통해 약하게 결합하도록 변경
