+++ src/main/java/nextstep/subway/member/application/MemberService.java
@@ -1,5 +1,6 @@
 package nextstep.subway.member.application;
 
+import nextstep.subway.error.SubwayException;
 import nextstep.subway.member.domain.Member;
 import nextstep.subway.member.domain.MemberRepository;
 import nextstep.subway.member.dto.MemberRequest;
@@ -23,7 +24,7 @@ public class MemberService {
     }
 
     public MemberResponse findMemberResponse(Long id) {
-        Member member = memberRepository.findById(id).orElseThrow(RuntimeException::new);
+        Member member = memberRepository.findById(id).orElseThrow(() -> new SubwayException("사용자가 없습니다."));
         return MemberResponse.of(member);
     }
